Separate the DartSass addon
